Move HTTPClient and implementations to standalone folders representing modules
