+++ EssentialFeedTests/Shared API Infra/URLSessionHTTPClientTests.swift
@@ -148,65 +148,3 @@ class URLSessionHTTPClientTests: XCTestCase {
         return URLResponse(url: makeURL(), mimeType: nil, expectedContentLength: 1, textEncodingName: nil)
     }
 }
-
-private class URLProtocolStub: URLProtocol {
-    private struct Stub {
-        let data: Data?
-        let response: URLResponse?
-        let error: Error?
-        let requestObserver: ((URLRequest) -> Void)?
-    }
-
-    private static var _stub: Stub?
-    private static var stub: Stub? {
-        get { return queue.sync { _stub } }
-        set { queue.sync { _stub = newValue } }
-    }
-
-    private static let queue = DispatchQueue(label: "URLProtocolStub.queue")
-
-    static func setStub(data: Data?, response: URLResponse?, error: Error?) {
-        URLProtocolStub.stub = Stub(data: data, response: response, error: error, requestObserver: nil)
-    }
-
-    static func observeRequest(observer: @escaping ((URLRequest) -> Void)) {
-        URLProtocolStub.stub = Stub(data: nil, response: nil, error: nil, requestObserver: observer)
-    }
-
-    static func startInterceptingRequests() {
-        URLProtocol.registerClass(URLProtocolStub.self)
-    }
-
-    static func stopInterceptingRequests() {
-        URLProtocol.unregisterClass(URLProtocolStub.self)
-        URLProtocolStub.stub = nil
-    }
-
-    override class func canInit(with request: URLRequest) -> Bool { return true }
-
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest { return request }
-
-    override func startLoading() {
-        guard let stub = URLProtocolStub.stub else { return }
-
-        if let requestObserver = stub.requestObserver {
-            return requestObserver(request)
-        }
-
-        if let data = stub.data {
-            client?.urlProtocol(self, didLoad: data)
-        }
-
-        if let response = stub.response {
-            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-        }
-
-        if let error = stub.error {
-            client?.urlProtocol(self, didFailWithError: error)
-        }
-
-        client?.urlProtocolDidFinishLoading(self)
-    }
-
-    override func stopLoading() {}
-}
